Definindo estrutura de pastas no front-end #49
