+++ client/src/App.css
@@ -5,118 +5,4 @@
   box-sizing: border-box;
 }
 
-.container-add-recepcionist, .container-search-recepcionist, .container-query-result {
-  display: flex;
-  width: 25vw;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  background-color: #E0DCDC;
-  border-radius: 8px;
-  margin-top: 100px;
-  padding: 40px 20px;
-}
 
-.container-add-recepcionist{
-  height: 400px;
-  margin-top: 40px;
-  margin-left: 100px;
-}
-
-.form-add-recepcionist, .form-search-recepcionist {
-  width: 70%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  
-}
-
-.container-add-recepcionist h1, .form-search-recepcionist h1 {
-  margin-bottom: 30px;
-  font-size: 24px;
-}
-
-.form-add-recepcionist input, .form-search-recepcionist input {
-  width: 100%;
-  height: 40px;
-  margin-bottom: 10px;
-  border-radius: 8px;
-  border: none;
-  text-indent: 20px;
-  outline: none;
-  padding: 5px;
-
-}
-
-.form-add-recepcionist button, .form-search-recepcionist button {
-  width: 100px;
-  height: 40px;
-  padding: 10px;
-  border-radius: 8px;
-  border: none;
-  outline: none;
-  margin-top: 10px;
-  color: #fff;
-  background-color: rgb(137, 163, 224);
-  cursor: pointer;
-  transition: 1s;
-}
-
-.form-add-recepcionist button:hover, .form-search-recepcionist button:hover {
-  background-color: #37BEB0;
-}
-
-.App {
-  display: flex;
-  justify-content: space-evenly;
-}
-
-.search-container {
-  margin-left: 50px;
-  margin-top: -60px;
-}
-
-#input-id {
-  display: none;
-}
-
-table {
-  margin-top: 30px;
-  width: 350px;
-  text-align: center;
-  border-collapse: collapse;
-  border: thin solid #c4c4c4;
-}
-
-td, th {
-  padding: 10px;
-}
-
-table thead tr th:first-of-type{
-  border-top-left-radius: 5px;
-}
-
-table thead tr th:last-of-type{
-  border-top-right-radius: 5px;
-}
-
-thead tr{
-  background-color: #b1a9a9;
-}
-
-thead {
-  border-radius: 16px;
-}
-
-#search-byId {
-  margin-top: 20px;
-}
-
-.label-get {
-  margin-bottom: 3px;
-}
-
-.error-user {
-  display: none;
-}
