+++ client/src/index.css
@@ -3,7 +3,9 @@
 
 body {
   margin: 0;
-  background-color: azure;
+ 
+  background-color: #E0DCDC;
+  
   font-family: 'Nunito', sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
